Split ledger tests into multiple test files/groups
